prothomalo_backend/requirements.txt
fastapi
uvicorn
elasticsearch
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
elastic-transport==8.17.1
elasticsearch==9.0.2
fastapi==0.115.13
Flask==3.1.1
flask-cors==6.0.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.3.1
pandas==2.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
Werkzeug==3.1.3
